requirements.txt
fastapi
uvicorn
psycopg2-binary
redis
selenium
pandas
matplotlib
seaborn